(maint) Add `rake table` task + update README.md
Before this patch, it wasn't clear which versions of each OS had data
for any given version of facter.

This patch addresses that problem by adding a **Facter version and
Operating System coverage** table to the `README.md`.  The markdown for
the table can be maintained using the new rake task, `rake table`.

Glenn Sarti: This commit is based on the work by op-ct