(maint) Add better support for Windows detection using source injection
Previously the fact information was not enough to determine the Windows version.
This is due to deficiencies in the facter facts not so much this gem.  This
commit adds a helper into the fact reading to record the filename used to load
the factsets. This is then used as a hint when generating the README table.
This is enabled using an environment variable, and is off by default.  This is
because these facts would not exist in a normal factset for a system of that
type.