(#78) Add tests for facter version
Previously many factset files were in the wrong directory which made maintenance
and debugging difficult.  This commit adds tests to ensure that the factset
files are indeed JSON parsable and that the facter version in the file matches
the directory it is located in.

This test is maked as skipped until all of the bad factset files can be changed.