Move the supported table out of README
This changes the workflow to have a rake task to create a database.md
file and then render that into the generated documentation.

The release process is also enhanced to include the generated
database.md file. This should allow sites like rubydoc.info correctly
render it as well, though this hasn't been tested.

It also introduces a GitHub Action workflow to build GitHub Pages with
the generated documentation. That means that the documentation should
always be up to date and no manual steps are needed anymore. It also
saves a lot of noise in the git log.