Remove Legacy Facts
Puppet 8 defaults to not providing Legacy Facts [1].  In order for our
tests to be relevant, we should not provide these facts.  We also had
special instructions to ask contributors to add these legacy facts which
are now obsolete.

As a baseline, do not provide any legacy facts.  Adjust references to
legacy fact to use non deprecated and still relevant ones in CI, and
provide a script that can asists us in updating the existing factsets.

[1]: https://github.com/puppetlabs/puppet/wiki/Puppet-8-Compatibility#legacy-facts